requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langgraph>=0.4.0
langgraph>=0.4.0